refactor : BoardTest 분리 - ChessGame로직 테스트를 GameTest로 분리
